+++ Assets/Scripts/PauseScreenManager.cs
@@ -2,6 +2,7 @@ using UnityEngine;
 using UnityEngine.Networking;
 using UnityEngine.SceneManagement;
 using System.Collections;
+using UnityEngine.Networking.Match;
 
 public class PauseScreenManager : MonoBehaviour {
     private bool opened = false;
@@ -41,6 +42,14 @@ public class PauseScreenManager : MonoBehaviour {
 
         SceneManager.LoadScene("MainMenu");
     }
+
+    public void Disconnect() {
+        NetworkManager nm = NetworkManager.singleton;
+        MatchInfo match = nm.matchInfo;
+        nm.matchMaker.DropConnection(match.networkId, match.nodeId, 0, nm.OnDropConnection);
+        nm.StopHost();
+    }
+
     public bool getOpened() {
         return opened;
     }