+++ Assets/Scripts/Player.cs
@@ -18,6 +18,10 @@ public class Player : NetworkBehaviour {
         sr = GetComponent<SpriteRenderer>();
         if (sr != null)
             sr.color = Color.blue;
+
+        //Set local camera to follow this player
+        SmoothCamera sc = GameObject.FindGameObjectWithTag("MainCamera").GetComponent<SmoothCamera>();
+        if (sc != null) sc.lookAt = this.transform;
     }
 
     public void Start() {